requirements.txt
google-ads
psycopg2-binary
python-dotenv
